+++ src/pacx/cli.py
@@ -11,6 +11,7 @@ from .cli import (
     pages,
     power_platform,
     profile,
+    pva,
     users,
 )
 
@@ -28,5 +29,6 @@ __all__ = [
     "pages",
     "power_platform",
     "profile",
+    "pva",
     "users",
 ]