+++ src/pacx/clients/__init__.py
@@ -2,6 +2,7 @@ from .authorization import AuthorizationRbacClient as AuthorizationRbacClient
 from .connectors import ConnectorsClient as ConnectorsClient
 from .dataverse import DataverseClient as DataverseClient
 from .power_platform import PowerPlatformClient as PowerPlatformClient
+from .pva import PVAClient as PVAClient
 from .user_management import (
     UserManagementClient as UserManagementClient,
     UserManagementOperationHandle as UserManagementOperationHandle,
@@ -12,6 +13,7 @@ __all__ = [
     "ConnectorsClient",
     "DataverseClient",
     "PowerPlatformClient",
+    "PVAClient",
     "UserManagementClient",
     "UserManagementOperationHandle",
 ]